+++ src/test/java/io/bitgrillr/gocddockerexecplugin/DockerExecPluginTest.java
@@ -3,6 +3,7 @@ package io.bitgrillr.gocddockerexecplugin;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.when;
@@ -32,7 +33,7 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({JobConsoleLogger.class, DockerUtils.class})
+@PrepareForTest({JobConsoleLogger.class, DockerUtils.class, SystemHelper.class})
 public class DockerExecPluginTest {
 
   @Test
@@ -108,13 +109,17 @@ public class DockerExecPluginTest {
 
   @Test
   public void handleExecute() throws Exception {
+    UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.pullImage("ubuntu:latest");
-    when(DockerUtils.createContainer("ubuntu:latest")).thenReturn("123");
-    when(DockerUtils.execCommand(eq("123"), any())).thenReturn(0);
+    DockerUtils.pullImage(anyString());
+    when(DockerUtils.createContainer(anyString(), anyString())).thenReturn("123");
+    when(DockerUtils.getContainerUid(anyString())).thenReturn("4:5");
+    when(DockerUtils.execCommand(anyString(), any(), any())).thenReturn(0);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.removeContainer("123");
+    DockerUtils.removeContainer(anyString());
+    PowerMockito.mockStatic(SystemHelper.class);
+    when(SystemHelper.getSystemUid()).thenReturn("7:8");
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -123,26 +128,54 @@ public class DockerExecPluginTest {
     image.put("value", "ubuntu:latest");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "pipelines/test");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);
 
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.pullImage("ubuntu:latest");
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.createContainer("ubuntu:latest",
+        Paths.get(System.getProperty("user.dir"), "pipelines/test").toAbsolutePath().toString());
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.getContainerUid("123");
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.execCommand("123", "root", "chown", "-R", "4:5", ".");
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.execCommand(eq("123"), eq(null), anyString(), any());
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.execCommand("123", "root", "chown", "-R", "7:8", ".");
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.removeContainer("123");
     assertEquals("Expected 2xx response", DefaultGoPluginApiResponse.SUCCESS_RESPONSE_CODE,
         response.responseCode());
     final JsonObject responseBody = Json.createReader(new StringReader(response.responseBody())).readObject();
     assertEquals("Expected success", Boolean.TRUE, responseBody.getBoolean("success"));
-    assertEquals("Wrong message", "Command 'cat /etc/os-release' completed with status 0",
+    assertEquals("Wrong message", "Command 'bash '-c' 'touch test && ls -l'' completed with status 0",
         Json.createReader(new StringReader(response.responseBody())).readObject().getString("message"));
   }
 
   @Test
   public void handleExecuteFailure() throws Exception {
+    UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.pullImage("ubuntu:latest");
-    when(DockerUtils.createContainer("ubuntu:latest")).thenReturn("123");
-    when(DockerUtils.execCommand(eq("123"), any())).thenReturn(1);
+    DockerUtils.pullImage(anyString());
+    when(DockerUtils.createContainer(anyString(), anyString())).thenReturn("123");
+    when(DockerUtils.getContainerUid(anyString())).thenReturn("4:5");
+    when(DockerUtils.execCommand(anyString(), any(), anyString(), any())).thenAnswer(i -> {
+      if (i.getArgument(2).equals("chown")) {
+        return 0;
+      } else {
+        return 1;
+      }
+    });
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.removeContainer("123");
+    DockerUtils.removeContainer(anyString());
+    PowerMockito.mockStatic(SystemHelper.class);
+    when(SystemHelper.getSystemUid()).thenReturn("7:8");
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -151,19 +184,27 @@ public class DockerExecPluginTest {
     image.put("value", "ubuntu:latest");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "/build");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);
 
     assertEquals("Expected 2xx response", DefaultGoPluginApiResponse.SUCCESS_RESPONSE_CODE, response.responseCode());
     final JsonObject responseBody = Json.createReader(new StringReader(response.responseBody())).readObject();
     assertEquals("Expected failure", Boolean.FALSE, responseBody.getBoolean("success"));
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.execCommand("123", "root", "chown", "-R", "4:5", ".");
+    PowerMockito.verifyStatic(DockerUtils.class);
+    DockerUtils.execCommand("123", "root", "chown", "-R", "7:8", ".");
   }
 
   @Test
   public void handleExecuteImageNotFound() throws Exception {
+    UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doThrow(new ImageNotFoundException("idont:exist")).when(DockerUtils.class);
-    DockerUtils.pullImage("idont:exist");
+    DockerUtils.pullImage(anyString());
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -172,6 +213,9 @@ public class DockerExecPluginTest {
     image.put("value", "idont:exist");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "/build");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);
 
@@ -186,11 +230,14 @@ public class DockerExecPluginTest {
     UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.pullImage("ubuntu:latest");
-    when(DockerUtils.createContainer("ubuntu:latest")).thenReturn("123");
-    when(DockerUtils.execCommand(eq("123"), any())).thenThrow(new DockerException("FAIL"));
+    DockerUtils.pullImage(anyString());
+    when(DockerUtils.createContainer(anyString(), anyString())).thenReturn("123");
+    when(DockerUtils.getContainerUid(anyString())).thenReturn("4:5");
+    when(DockerUtils.execCommand(anyString(), any(), anyString(), any())).thenThrow(new DockerException("FAIL"));
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.removeContainer("123");
+    DockerUtils.removeContainer(anyString());
+    PowerMockito.mockStatic(SystemHelper.class);
+    when(SystemHelper.getSystemUid()).thenReturn("7:8");
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -199,6 +246,9 @@ public class DockerExecPluginTest {
     image.put("value", "ubuntu:latest");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "/build");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);
 
@@ -213,11 +263,14 @@ public class DockerExecPluginTest {
     UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.pullImage("ubuntu:latest");
-    when(DockerUtils.createContainer("ubuntu:latest")).thenReturn("123");
-    when(DockerUtils.execCommand(eq("123"), any())).thenReturn(0);
+    DockerUtils.pullImage(anyString());
+    when(DockerUtils.createContainer(anyString(), anyString())).thenReturn("123");
+    when(DockerUtils.getContainerUid(anyString())).thenReturn("4:5");
+    when(DockerUtils.execCommand(anyString(), any(), anyString(), any())).thenReturn(0);
     PowerMockito.doThrow(new DockerException("FAIL")).when(DockerUtils.class);
-    DockerUtils.removeContainer("123");
+    DockerUtils.removeContainer(anyString());
+    PowerMockito.mockStatic(SystemHelper.class);
+    when(SystemHelper.getSystemUid()).thenReturn("7:8");
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -226,6 +279,9 @@ public class DockerExecPluginTest {
     image.put("value", "ubuntu:latest");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "/build");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);
 
@@ -240,11 +296,14 @@ public class DockerExecPluginTest {
     UnitTestUtils.mockJobConsoleLogger();
     PowerMockito.mockStatic(DockerUtils.class);
     PowerMockito.doNothing().when(DockerUtils.class);
-    DockerUtils.pullImage("ubuntu:latest");
-    when(DockerUtils.createContainer("ubuntu:latest")).thenReturn("123");
-    when(DockerUtils.execCommand(eq("123"), any())).thenThrow(new DockerException("FAIL1"));
+    DockerUtils.pullImage(anyString());
+    when(DockerUtils.createContainer(anyString(), anyString())).thenReturn("123");
+    when(DockerUtils.getContainerUid(anyString())).thenReturn("4:5");
+    when(DockerUtils.execCommand(anyString(), any(), anyString(), any())).thenThrow(new DockerException("FAIL1"));
     PowerMockito.doThrow(new DockerException("FAIL2")).when(DockerUtils.class);
-    DockerUtils.removeContainer("123");
+    DockerUtils.removeContainer(anyString());
+    PowerMockito.mockStatic(SystemHelper.class);
+    when(SystemHelper.getSystemUid()).thenReturn("7:8");
 
     DefaultGoPluginApiRequest request = new DefaultGoPluginApiRequest(null, null, "execute");
     Map<String, Object> requestBody = new HashMap<>();
@@ -253,6 +312,9 @@ public class DockerExecPluginTest {
     image.put("value", "ubuntu:latest");
     config.put("IMAGE", image);
     requestBody.put("config", config);
+    Map<String, Object> context = new HashMap<>();
+    context.put("workingDirectory", "/build");
+    requestBody.put("context", context);
     request.setRequestBody(Json.createObjectBuilder(requestBody).build().toString());
     final GoPluginApiResponse response = new DockerExecPlugin().handle(request);